feat(secureboot): expose genkey as pkg/secureboot.GenerateKeySet
Move the Secure Boot key generation logic out of internal/cmd/genkey.go
into a new public package pkg/secureboot so it can be consumed as a
Go library without shelling out to the auroraboot binary.

The existing 'auroraboot genkey' CLI command is preserved as a thin
wrapper around secureboot.GenerateKeySet — same flags, same behavior.

External tools (e.g. daedalus) can now call:

    import "github.com/kairos-io/AuroraBoot/pkg/secureboot"

    err := secureboot.GenerateKeySet(secureboot.Options{
        Name:      "production",
        OutputDir: "/path/to/keys",
    })

The pipeline still relies on openssl being available on PATH for X.509
certificate creation; .auth/.esl files are produced via foxboron/sbctl
and foxboron/go-uefi (no external commands).
